+++ src/Namicorn.test.ts
@@ -1,5 +1,5 @@
 import nock from 'nock';
-import Namicorn from './namicorn';
+import Namicorn from '.';
 import _ from 'lodash';
 import mockData from './testData/mockData.json';
 import Ens from './ens';
@@ -7,6 +7,11 @@ import Ens from './ens';
 const DefaultUrl = 'https://unstoppabledomains.com/api/v1';
 const MainnetUrl = 'https://mainnet.infura.io';
 const ZilliqaUrl = 'https://api.zilliqa.com';
+
+const EnsRegexUrlPattern = new RegExp(/(https:\/\/)(\w+)(.infura.io)\/?/);
+const ZnsRegexUrlPattern = new RegExp(
+  /(https:\/\/)((dev-)?\w+)(.zilliqa.com)\/?/,
+);
 
 const mockAPICalls = (topLevel: string, testName: string, url = MainnetUrl) => {
   if (process.env.LIVE) {
@@ -73,13 +78,14 @@ describe('ZNS', () => {
 
   it("doesn't support zil domain when zns is disabled", () => {
     const namicorn = new Namicorn({ blockchain: { zns: false } });
+    expect(namicorn.zns).toBeUndefined();
     expect(namicorn.isSupportedDomain('hello.zil')).toBeFalsy();
   });
 
   it('checks normalizeSource zns (boolean)', async () => {
     const namicorn = new Namicorn({ blockchain: { zns: true } });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (boolean - false)', async () => {
@@ -93,7 +99,7 @@ describe('ZNS', () => {
       blockchain: { zns: 'https://api.zilliqa.com' },
     });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns wrong string', async () => {
@@ -107,13 +113,14 @@ describe('ZNS', () => {
       blockchain: { zns: { url: 'https://api.zilliqa.com' } },
     });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (object) #2', async () => {
-    expect(
-      () => new Namicorn({ blockchain: { zns: { network: 333 } } }),
-    ).toThrow();
+    const namicorn = new Namicorn({ blockchain: { zns: { network: 333 } } });
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
+    expect(namicorn.zns.network).toBe('testnet');
+    expect(namicorn.zns.registryAddress).toBeUndefined();
   });
 
   it('checks normalizeSource zns (object) #3', async () => {
@@ -121,7 +128,7 @@ describe('ZNS', () => {
       blockchain: { zns: { url: 'https://api.zilliqa.com' } },
     });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (object) #4', async () => {
@@ -129,16 +136,17 @@ describe('ZNS', () => {
       blockchain: { zns: { url: 'https://api.zilliqa.com', network: 1 } },
     });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (object) #5', async () => {
-    expect(
-      () =>
-        new Namicorn({
-          blockchain: { zns: { url: 'https://api.zilliqa.com', network: 333 } },
-        }),
-    ).toThrow();
+    const namicorn = new Namicorn({
+      blockchain: { zns: { url: 'https://api.zilliqa.com', network: 333 } },
+    });
+
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
+    expect(namicorn.zns.network).toBe('testnet');
+    expect(namicorn.zns.registryAddress).toBeUndefined();
   });
 
   it('checks normalizeSource zns (object) #6', async () => {
@@ -158,16 +166,17 @@ describe('ZNS', () => {
       blockchain: { zns: { network: 'mainnet' } },
     });
     expect(namicorn.zns.network).toBe('mainnet');
-    expect(namicorn.zns.url).toBe('https://api.zilliqa.com');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (object) #9', async () => {
-    expect(
-      () =>
-        new Namicorn({
-          blockchain: { zns: { network: 'testnet' } },
-        }),
-    ).toThrow();
+    const namicorn = new Namicorn({
+      blockchain: { zns: { network: 'testnet' } },
+    });
+
+    expect(namicorn.zns.network).toBe('testnet');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
+    expect(namicorn.zns.registryAddress).toBeUndefined();
   });
 
   it('checks normalizeSource zns (object) #10', async () => {
@@ -180,18 +189,21 @@ describe('ZNS', () => {
     expect(namicorn.zns.registryAddress).toBe(
       'zil1jcgu2wlx6xejqk9jw3aaankw6lsjzeunx2j0jz',
     );
-    expect(namicorn.zns.url).toBe(ZilliqaUrl);
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
   });
 
   it('checks normalizeSource zns (object) #11', async () => {
-    expect(
-      () =>
-        new Namicorn({
-          blockchain: {
-            zns: { registry: '0xabcffff1231586348194fcabbeff1231240234fc' },
-          },
-        }),
-    ).toThrow();
+    const namicorn = new Namicorn({
+      blockchain: {
+        zns: { registry: '0xabcffff1231586348194fcabbeff1231240234fc' },
+      },
+    });
+
+    expect(namicorn.zns.network).toBe('mainnet');
+    expect(namicorn.zns.url).toMatch(ZnsRegexUrlPattern);
+    expect(namicorn.zns.registryAddress).toBe(
+      '0xabcffff1231586348194fcabbeff1231240234fc',
+    );
   });
 });
 
@@ -203,7 +215,7 @@ describe('ENS', () => {
     const namicorn = new Namicorn({
       blockchain: { ens: { network: 'mainnet' } },
     });
-    expect(namicorn.ens.url).toEqual('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
     expect(namicorn.ens.network).toEqual('mainnet');
   });
 
@@ -214,7 +226,7 @@ describe('ENS', () => {
     const namicorn = new Namicorn({
       blockchain: { ens: true },
     });
-    expect(namicorn.ens.url).toEqual('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
     expect(namicorn.ens.network).toEqual('mainnet');
     var result = await namicorn.address('matthewgould.eth', 'ETH');
     expect(result).toEqual('0x714ef33943d925731FBB89C99aF5780D888bD106');
@@ -339,14 +351,13 @@ describe('ENS', () => {
   it('checks if the network is supported(false)', async () => {
     const ens = new Ens({ network: 5 });
     const answer = ens.isSupportedNetwork();
-    const testnew = new Namicorn({ blockchain: { ens: { network: 1 } } });
     expect(answer).toBe(false);
   });
 
   it('checks normalizeSource ens (boolean)', async () => {
     const namicorn = new Namicorn({ blockchain: { ens: true } });
     expect(namicorn.ens.network).toBe('mainnet');
-    expect(namicorn.ens.url).toBe('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (boolean - false)', async () => {
@@ -361,13 +372,13 @@ describe('ENS', () => {
       blockchain: { ens: { url: 'https://mainnet.infura.io' } },
     });
     expect(namicorn.ens.network).toBe('mainnet');
-    expect(namicorn.ens.url).toBe('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #2', async () => {
     const namicorn = new Namicorn({ blockchain: { ens: { network: 3 } } });
     expect(namicorn.ens.network).toBe('ropsten');
-    expect(namicorn.ens.url).toBe('https://ropsten.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #3', async () => {
@@ -375,7 +386,7 @@ describe('ENS', () => {
       blockchain: { ens: { url: 'https://rinkeby.infura.io' } },
     });
     expect(namicorn.ens.network).toBe('rinkeby');
-    expect(namicorn.ens.url).toBe('https://rinkeby.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #4', async () => {
@@ -383,7 +394,7 @@ describe('ENS', () => {
       blockchain: { ens: { url: 'https://goerli.infura.io', network: 5 } },
     });
     expect(namicorn.ens.network).toBe('goerli');
-    expect(namicorn.ens.url).toBe('https://goerli.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #6', async () => {
@@ -407,7 +418,7 @@ describe('ENS', () => {
       blockchain: { ens: { network: 'mainnet' } },
     });
     expect(namicorn.ens.network).toBe('mainnet');
-    expect(namicorn.ens.url).toBe('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #9', async () => {
@@ -415,7 +426,7 @@ describe('ENS', () => {
       blockchain: { ens: { network: 'kovan' } },
     });
     expect(namicorn.ens.network).toBe('kovan');
-    expect(namicorn.ens.url).toBe('https://kovan.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
   });
 
   it('checks normalizeSource ens (object) #10', async () => {
@@ -425,7 +436,7 @@ describe('ENS', () => {
       },
     });
     expect(namicorn.ens.network).toBe('mainnet');
-    expect(namicorn.ens.url).toBe('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
     expect(namicorn.ens.registryAddress).toBe(
       '0x314159265dd8dbb310642f98f50c066173c1259b',
     );
@@ -441,7 +452,7 @@ describe('ENS', () => {
       },
     });
     expect(namicorn.ens.network).toBe('ropsten');
-    expect(namicorn.ens.url).toBe('https://ropsten.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
     expect(namicorn.ens.registryAddress).toBe(
       '0x112234455c3a32fd11230c42e7bccd4a84e02010',
     );
@@ -455,7 +466,7 @@ describe('ENS', () => {
     });
 
     expect(namicorn.ens.network).toBe('mainnet');
-    expect(namicorn.ens.url).toBe('https://mainnet.infura.io');
+    expect(namicorn.ens.url).toMatch(EnsRegexUrlPattern);
     expect(namicorn.ens.registryAddress).toBe(
       '0xabcffff1231586348194fcabbeff1231240234fc',
     );