+++ src/index.ts
@@ -1,2 +1,2 @@
-import Namicorn from './namicorn';
+import Namicorn from './Namicorn';
 export { Namicorn, Namicorn as default };